Move 'convenience' implementations into the peripheral interface
