+++ Sources/CombineCoreBluetooth/Models/PeripheralError.swift
@@ -1,0 +1,7 @@
+import Foundation
+import CoreBluetooth
+
+public enum PeripheralError: Error, Equatable {
+  case serviceNotFound(CBUUID)
+  case characteristicNotFound(CBUUID)
+}
